feat: refactorizar rutas de gestión de usuarios y moderación
- Actualizado enlace de AdminProfilePage para apuntar a usuarios de moderación.
- Eliminados componentes BlockUserModal y UsersManagementClient.
- Eliminada página de gestión de usuarios y pruebas asociadas.
- Mejoradas rutas de API de moderación para filtrar albergues y proveedores por estado.
- Actualizada API de bloqueo de usuarios para reflejar nuevas rutas de moderación.
- Ajustado AdminDashboardClient para eliminar enlace de gestión de usuarios.
- Refactorizado BlockUserButton para usar nueva ruta de modal.
- Actualizado EditUserButton para reflejar nuevas rutas de moderación.
- Modificadas constantes para incluir nuevos enlaces de navegación de moderación.
- Mejorado servicio de moderación para gestionar albergues y proveedores aprobados y rechazados.
