+++ app/(dashboard)/admin/moderation/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function ModerationPage() {
-  redirect("/admin/moderation/shelters");
+  redirect("/admin/moderation/users");
 }
