+++ app/(dashboard)/admin/moderation/vendors/vendor-moderation-client.tsx
@@ -26,6 +26,7 @@ interface Vendor {
   municipality: string;
   address: string;
   description: string | null;
+  rejectionReason?: string | null;
   createdAt: string;
   user: VendorUser;
 }
@@ -36,9 +37,15 @@ export function VendorModerationClient() {
   const [processingId, setProcessingId] = useState<string | null>(null);
   const [rejectModalData, setRejectModalData] = useState<{ id: string; name: string } | null>(null);
 
-  const fetchVendors = async () => {
+  const [activeTab, setActiveTab] = useState<"pending" | "approved" | "rejected">("pending");
+
+  const fetchVendors = async (tab: "pending" | "approved" | "rejected") => {
+    setLoading(true);
     try {
-      const res = await fetch("/api/admin/moderation/vendors");
+      const url = tab === "pending" 
+        ? "/api/admin/moderation/vendors" 
+        : `/api/admin/moderation/vendors?status=${tab}`;
+      const res = await fetch(url);
       if (!res.ok) throw new Error("Error al cargar solicitudes");
       const data = await res.json();
       setVendors(data);
@@ -50,8 +57,8 @@ export function VendorModerationClient() {
   };
 
   useEffect(() => {
-    fetchVendors();
-  }, []);
+    fetchVendors(activeTab);
+  }, [activeTab]);
 
   const handleApprove = async (id: string) => {
     setProcessingId(id);
@@ -105,10 +112,47 @@ export function VendorModerationClient() {
 
   return (
     <div className="space-y-6">
+      <div className="flex justify-center mb-6">
+        <div className="bg-gray-100 p-1 rounded-lg inline-flex gap-1 border border-gray-200">
+          <button
+            onClick={() => setActiveTab("pending")}
+            className={`px-4 py-2 text-sm font-medium rounded-md transition-all ${
+              activeTab === "pending"
+                ? "bg-white text-gray-900 shadow-sm"
+                : "text-gray-500 hover:text-gray-700 hover:bg-gray-200/50"
+            }`}
+          >
+            Pendientes
+          </button>
+          <button
+            onClick={() => setActiveTab("approved")}
+            className={`px-4 py-2 text-sm font-medium rounded-md transition-all ${
+              activeTab === "approved"
+                ? "bg-white text-gray-900 shadow-sm"
+                : "text-gray-500 hover:text-gray-700 hover:bg-gray-200/50"
+            }`}
+          >
+            Aprobados
+          </button>
+          <button
+            onClick={() => setActiveTab("rejected")}
+            className={`px-4 py-2 text-sm font-medium rounded-md transition-all ${
+              activeTab === "rejected"
+                ? "bg-white text-gray-900 shadow-sm"
+                : "text-gray-500 hover:text-gray-700 hover:bg-gray-200/50"
+            }`}
+          >
+            Rechazados
+          </button>
+        </div>
+      </div>
+
       <div className="flex justify-between items-center mb-6">
-        <h2 className="text-xl font-bold text-gray-900">Negocios Pendientes</h2>
+        <h2 className="text-xl font-bold text-gray-900">
+          {activeTab === "pending" ? "Negocios Pendientes" : activeTab === "approved" ? "Negocios Aprobados" : "Negocios Rechazados"}
+        </h2>
         <span className="px-3 py-1 bg-purple-100 text-purple-800 rounded-full text-sm font-semibold">
-          {vendors.length} solicitudes
+          {vendors.length} {activeTab === "pending" ? "solicitudes" : activeTab === "approved" ? "verificados" : "rechazados"}
         </span>
       </div>
 
@@ -118,7 +162,13 @@ export function VendorModerationClient() {
             <CheckCircle className="h-10 w-10 text-green-600" />
           </div>
           <h3 className="text-xl font-bold text-gray-900 mb-2">Todo al día</h3>
-          <p className="text-gray-500 max-w-sm">No hay solicitudes de nuevos negocios pendientes por revisar en este momento.</p>
+          <p className="text-gray-500 max-w-sm">
+            {activeTab === "pending" 
+              ? "No hay solicitudes de nuevos negocios pendientes por revisar en este momento." 
+              : activeTab === "approved" 
+                ? "No hay negocios aprobados registrados en el sistema."
+                : "No hay negocios rechazados en el sistema."}
+          </p>
         </div>
       ) : (
         <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
@@ -165,25 +215,44 @@ export function VendorModerationClient() {
                 )}
               </div>
               
-              <div className="p-4 bg-gray-50 flex gap-3">
-                <Button 
-                  className="flex-1 bg-white border-red-200 text-red-600 hover:bg-red-50 hover:text-red-700 transition" 
-                  variant="outline"
-                  onClick={() => setRejectModalData({ id: vendor.id, name: vendor.businessName })}
-                  disabled={processingId === vendor.id}
-                >
-                  <X className="h-4 w-4 mr-2" />
-                  Rechazar
-                </Button>
-                <Button 
-                  className="flex-1 bg-green-600 hover:bg-green-700 text-white transition"
-                  onClick={() => handleApprove(vendor.id)}
-                  disabled={processingId === vendor.id}
-                >
-                  {processingId === vendor.id ? <Loader2 className="h-4 w-4 animate-spin mr-2" /> : <Check className="h-4 w-4 mr-2" />}
-                  Aprobar
-                </Button>
-              </div>
+              {activeTab === "pending" ? (
+                <div className="p-4 bg-gray-50 flex gap-3">
+                  <Button 
+                    className="flex-1 bg-white border-red-200 text-red-600 hover:bg-red-50 hover:text-red-700 transition" 
+                    variant="outline"
+                    onClick={() => setRejectModalData({ id: vendor.id, name: vendor.businessName })}
+                    disabled={processingId === vendor.id}
+                  >
+                    <X className="h-4 w-4 mr-2" />
+                    Rechazar
+                  </Button>
+                  <Button 
+                    className="flex-1 bg-green-600 hover:bg-green-700 text-white transition"
+                    onClick={() => handleApprove(vendor.id)}
+                    disabled={processingId === vendor.id}
+                  >
+                    {processingId === vendor.id ? <Loader2 className="h-4 w-4 animate-spin mr-2" /> : <Check className="h-4 w-4 mr-2" />}
+                    Aprobar
+                  </Button>
+                </div>
+              ) : activeTab === "approved" ? (
+                <div className="p-4 bg-green-50 flex items-center justify-center text-green-700 font-medium border-t border-green-100">
+                  <CheckCircle className="h-5 w-5 mr-2" />
+                  Negocio Verificado
+                </div>
+              ) : (
+                <div className="p-4 bg-red-50 flex flex-col justify-center text-red-700 font-medium border-t border-red-100 text-sm">
+                  <div className="flex items-center mb-1">
+                    <X className="h-5 w-5 mr-2" />
+                    <span>Solicitud Rechazada</span>
+                  </div>
+                  {vendor.rejectionReason && (
+                    <p className="text-xs text-red-600 mt-1 italic">
+                      Motivo: {vendor.rejectionReason}
+                    </p>
+                  )}
+                </div>
+              )}
             </div>
           ))}
         </div>
