+++ app/(dashboard)/admin/profile/page.tsx
@@ -25,7 +25,7 @@ export default async function AdminProfilePage() {
   return (
     <main className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
       <div className="mb-8">
-        <Link href="/admin/users" className="inline-flex items-center gap-2 text-purple-600 hover:text-purple-700 text-sm font-semibold">
+        <Link href="/admin/moderation/users" className="inline-flex items-center gap-2 text-purple-600 hover:text-purple-700 text-sm font-semibold">
           <ArrowLeft className="w-4 h-4" />
           Volver a Usuarios
         </Link>
